@@ -11,8 +11,6 @@ Devise.setup do |config|
     helper :url
   end
 
-  config.use_salt_as_remember_token = true
-
   # Automatically apply schema changes in tableless databases
  # config.apply_schema = false
 
@@ -207,17 +205,5 @@ Devise.setup do |config|
   # The default HTTP method used to sign out a resource. Default is :delete.
   config.sign_out_via = :delete
 
-  # ==> OmniAuth
-  # Add a new OmniAuth provider. Check the wiki for more information on setting
-  # up on your models and hooks.
-  # config.omniauth :github, 'APP_ID', 'APP_SECRET', :scope => 'user,public_repo'
-
-  # ==> Warden configuration
-  # If you want to use other strategies, that are not supported by Devise, or
-  # change the failure app, you can configure them inside the config.warden block.
-  #
-  # config.warden do |manager|
-  #   manager.intercept_401 = false
-  #   manager.default_strategies(:scope => :user).unshift :some_external_strategy
-  # end
+  config.secret_key = DEVISE_SECRET_KEY
 end
